(1) Condensed DeckBase, DeckDraw, DeckDiscard into a single "Deck" class and added a "location" attribute to cards that will indicate where a card currently resides.
(2) Added a new unocardlocation enumeration to globals to support the new location attribute used by cards.

(3) Created a Saver class that can serialize and deserialize any class (even nested classes and complex data structures) using a combination of FullSerializer (to convert back and forth to a JSON string), and Unity's PlayerPrefs class (which persists data in a platform-neutral way and works on every platfrom we are targeting).

(4) Added the FullSerializer DLL to the Plugins folder.  The Plugins folder is "special" to Unity -- any DLLs that are there will be compiled in such a way that they are always loaded first before the application executes.

(5) Added several tests to the Deck class and the Saver class to test Serialization and Deserialization. 

All tests successful and look correct.